#27 - Part the files in folders - MenuPage
